pallet_admin_utils::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

schedule_grandpa_change

A public interface for pallet_grandpa::Pallet::schedule_grandpa_change.

sudo_set_activity_cutoff

The extrinsic sets the activity cutoff for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the activity cutoff.

sudo_set_adjustment_alpha

The extrinsic sets the adjustment alpha for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the adjustment alpha.

sudo_set_adjustment_interval

The extrinsic sets the adjustment interval for a subnet. It is only callable by the root account, not changeable by the subnet owner. The extrinsic will call the Subtensor pallet to set the adjustment interval.

sudo_set_alpha_sigmoid_steepness

Arguments

sudo_set_alpha_values

Sets values for liquid alpha

sudo_set_bonds_moving_average

The extrinsic sets the bonds moving average for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the bonds moving average.

sudo_set_bonds_penalty

The extrinsic sets the bonds penalty for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the bonds penalty.

sudo_set_coldkey_swap_schedule_duration

Sets the duration of the coldkey swap schedule.

sudo_set_commit_reveal_weights_enabled

The extrinsic enabled/disables commit/reaveal for a given subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the value.

sudo_set_commit_reveal_weights_interval

Sets the commit-reveal weights periods for a specific subnet.

sudo_set_default_take

The extrinsic sets the default take for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the default take.

sudo_set_difficulty

The extrinsic sets the difficulty for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the difficulty.

sudo_set_dissolve_network_schedule_duration

Sets the duration of the dissolve network schedule.

sudo_set_ema_price_halving_period

Arguments

sudo_set_evm_chain_id

Sets the EVM ChainID.

sudo_set_immunity_period

The extrinsic sets the immunity period for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the immunity period.

sudo_set_kappa

The extrinsic sets the kappa for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the kappa.

sudo_set_liquid_alpha_enabled

Enables or disables Liquid Alpha for a given subnet.

sudo_set_lock_reduction_interval

The extrinsic sets the lock reduction interval for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the lock reduction interval.

sudo_set_max_allowed_uids

The extrinsic sets the maximum allowed UIDs for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the maximum allowed UIDs for a subnet.

sudo_set_max_allowed_validators

The extrinsic sets the maximum allowed validators for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the maximum allowed validators.

sudo_set_max_burn

The extrinsic sets the maximum burn for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the maximum burn.

sudo_set_max_difficulty

The extrinsic sets the maximum difficulty for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the maximum difficulty.

sudo_set_max_registrations_per_block

The extrinsic sets the maximum registrations per block for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the maximum registrations per block.

sudo_set_max_weight_limit

The extrinsic sets the adjustment beta for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the adjustment beta.

sudo_set_min_allowed_weights

The extrinsic sets the minimum allowed weights for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the minimum allowed weights.

sudo_set_min_burn

The extrinsic sets the minimum burn for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the minimum burn.

sudo_set_min_delegate_take

The extrinsic sets the minimum delegate take. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the minimum delegate take.

sudo_set_min_difficulty

The extrinsic sets the minimum difficulty for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the minimum difficulty.

sudo_set_network_immunity_period

The extrinsic sets the immunity period for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the immunity period for the network.

sudo_set_network_max_stake

Sets the maximum stake allowed for a specific network.

sudo_set_network_min_lock_cost

The extrinsic sets the min lock cost for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the min lock cost for the network.

sudo_set_network_pow_registration_allowed

The extrinsic sets the network PoW registration allowed for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the network PoW registration allowed.

sudo_set_network_rate_limit

The extrinsic sets the network rate limit for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the network rate limit.

sudo_set_network_registration_allowed

The extrinsic sets the network registration allowed for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the network registration allowed.

sudo_set_nominator_min_required_stake

The extrinsic sets the minimum stake required for nominators. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the minimum stake required for nominators.

sudo_set_rao_recycled

The extrinsic sets the recycled RAO for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the recycled RAO.

sudo_set_rho

The extrinsic sets the rho for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the rho.

sudo_set_serving_rate_limit

The extrinsic sets the serving rate limit for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the serving rate limit.

sudo_set_sn_owner_hotkey

Sets or updates the hotkey account associated with the owner of a specific subnet.

sudo_set_stake_threshold

The extrinsic sets the weights min stake. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the weights min stake.

sudo_set_subnet_limit

The extrinsic sets the subnet limit for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the subnet limit.

sudo_set_subnet_moving_alpha

Arguments

sudo_set_subnet_owner_cut

The extrinsic sets the subnet owner cut for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the subnet owner cut.

sudo_set_subnet_owner_hotkey

Change the SubnetOwnerHotkey for a given subnet.

sudo_set_subtoken_enabled

Enables or disables subtoken trading for a given subnet.

sudo_set_target_registrations_per_interval

The extrinsic sets the target registrations per interval for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the target registrations per interval.

sudo_set_tempo

The extrinsic sets the tempo for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the tempo.

sudo_set_toggle_transfer

Enables or disables Liquid Alpha for a given subnet.

sudo_set_total_issuance

The extrinsic sets the total issuance for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the issuance for the network.

sudo_set_tx_delegate_take_rate_limit

The extrinsic sets the rate limit for delegate take transactions. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the rate limit for delegate take transactions.

sudo_set_tx_rate_limit

The extrinsic sets the transaction rate limit for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the transaction rate limit.

sudo_set_weights_set_rate_limit

The extrinsic sets the weights set rate limit for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the weights set rate limit.

sudo_set_weights_version_key

The extrinsic sets the weights version key for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the weights version key.

sudo_set_yuma3_enabled

Enables or disables Yuma3 for a given subnet.

sudo_toggle_evm_precompile

Toggles the enablement of an EVM precompile.

swap_authorities

The extrinsic sets the new authorities for Aura consensus. It is only callable by the root account. The extrinsic will call the Aura pallet to change the authorities.