pallet_admin_utils::pallet::dispatchables

Function schedule_grandpa_change

pub fn schedule_grandpa_change<T: Config>(
    next_authorities: AuthorityList,
    in_blocks: BlockNumberFor<T>,
    forced: Option<BlockNumberFor<T>>,
)

A public interface for pallet_grandpa::Pallet::schedule_grandpa_change.

Schedule a change in the authorities.

The change will be applied at the end of execution of the block in_blocks after the current block. This value may be 0, in which case the change is applied at the end of the current block.

If the forced parameter is defined, this indicates that the current set has been synchronously determined to be offline and that after in_blocks the given change should be applied. The given block number indicates the median last finalized block number and it should be used as the canon block when starting the new grandpa voter.

No change should be signaled while any change is pending. Returns an error if a change is already pending.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::schedule_grandpa_change.