pallet_admin_utils::pallet::dispatchables

Function sudo_set_bonds_moving_average

pub fn sudo_set_bonds_moving_average<T: Config>(
    netuid: u16,
    bonds_moving_average: u64,
)

The extrinsic sets the bonds moving average for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the bonds moving average.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::sudo_set_bonds_moving_average.