pallet_admin_utils::pallet::dispatchables

Function sudo_set_network_max_stake

pub fn sudo_set_network_max_stake<T: Config>(_netuid: u16, _max_stake: u64)

Sets the maximum stake allowed for a specific network.

This function allows the root account to set the maximum stake for a given network. It updates the network’s maximum stake value and logs the change.

Arguments

origin - The origin of the call, which must be the root account.

netuid - The unique identifier of the network.

max_stake - The new maximum stake value to set.

Returns

Returns Ok(()) if the operation is successful, or an error if it fails.

Example

Notes

This function can only be called by the root account.

The netuid should correspond to an existing network.

TODO

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::sudo_set_network_max_stake.