pallet_admin_utils::pallet::dispatchables

Function sudo_set_toggle_transfer

pub fn sudo_set_toggle_transfer<T: Config>(netuid: u16, toggle: bool)

Enables or disables Liquid Alpha for a given subnet.

Parameters

origin: The origin of the call, which must be the root account or subnet owner.

netuid: The unique identifier for the subnet.

enabled: A boolean flag to enable or disable Liquid Alpha.

Weight

This function has a fixed weight of 0 and is classified as an operational transaction that does not incur any fees.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::sudo_set_toggle_transfer.