pallet_admin_utils::pallet

Enum Call

pub enum Call<T: Config> {

    swap_authorities {
        new_authorities: BoundedVec<<T as Config>::AuthorityId, T::MaxAuthorities>,
    },
    sudo_set_default_take {
        default_take: u16,
    },
    sudo_set_tx_rate_limit {
        tx_rate_limit: u64,
    },
    sudo_set_serving_rate_limit {
        netuid: u16,
        serving_rate_limit: u64,
    },
    sudo_set_min_difficulty {
        netuid: u16,
        min_difficulty: u64,
    },
    sudo_set_max_difficulty {
        netuid: u16,
        max_difficulty: u64,
    },
    sudo_set_weights_version_key {
        netuid: u16,
        weights_version_key: u64,
    },
    sudo_set_weights_set_rate_limit {
        netuid: u16,
        weights_set_rate_limit: u64,
    },
    sudo_set_adjustment_interval {
        netuid: u16,
        adjustment_interval: u16,
    },
    sudo_set_adjustment_alpha {
        netuid: u16,
        adjustment_alpha: u64,
    },
    sudo_set_max_weight_limit {
        netuid: u16,
        max_weight_limit: u16,
    },
    sudo_set_immunity_period {
        netuid: u16,
        immunity_period: u16,
    },
    sudo_set_min_allowed_weights {
        netuid: u16,
        min_allowed_weights: u16,
    },
    sudo_set_max_allowed_uids {
        netuid: u16,
        max_allowed_uids: u16,
    },
    sudo_set_kappa {
        netuid: u16,
        kappa: u16,
    },
    sudo_set_rho {
        netuid: u16,
        rho: u16,
    },
    sudo_set_activity_cutoff {
        netuid: u16,
        activity_cutoff: u16,
    },
    sudo_set_network_registration_allowed {
        netuid: u16,
        registration_allowed: bool,
    },
    sudo_set_network_pow_registration_allowed {
        netuid: u16,
        registration_allowed: bool,
    },
    sudo_set_target_registrations_per_interval {
        netuid: u16,
        target_registrations_per_interval: u16,
    },
    sudo_set_min_burn {
        netuid: u16,
        min_burn: u64,
    },
    sudo_set_max_burn {
        netuid: u16,
        max_burn: u64,
    },
    sudo_set_difficulty {
        netuid: u16,
        difficulty: u64,
    },
    sudo_set_max_allowed_validators {
        netuid: u16,
        max_allowed_validators: u16,
    },
    sudo_set_bonds_moving_average {
        netuid: u16,
        bonds_moving_average: u64,
    },
    sudo_set_bonds_penalty {
        netuid: u16,
        bonds_penalty: u16,
    },
    sudo_set_max_registrations_per_block {
        netuid: u16,
        max_registrations_per_block: u16,
    },
    sudo_set_subnet_owner_cut {
        subnet_owner_cut: u16,
    },
    sudo_set_network_rate_limit {
        rate_limit: u64,
    },
    sudo_set_tempo {
        netuid: u16,
        tempo: u16,
    },
    sudo_set_total_issuance {
        total_issuance: u64,
    },
    sudo_set_network_immunity_period {
        immunity_period: u64,
    },
    sudo_set_network_min_lock_cost {
        lock_cost: u64,
    },
    sudo_set_subnet_limit {
        max_subnets: u16,
    },
    sudo_set_lock_reduction_interval {
        interval: u64,
    },
    sudo_set_rao_recycled {
        netuid: u16,
        rao_recycled: u64,
    },
    sudo_set_stake_threshold {
        min_stake: u64,
    },
    sudo_set_nominator_min_required_stake {
        min_stake: u64,
    },
    sudo_set_tx_delegate_take_rate_limit {
        tx_rate_limit: u64,
    },
    sudo_set_min_delegate_take {
        take: u16,
    },
    sudo_set_commit_reveal_weights_enabled {
        netuid: u16,
        enabled: bool,
    },
    sudo_set_liquid_alpha_enabled {
        netuid: u16,
        enabled: bool,
    },
    sudo_set_alpha_values {
        netuid: u16,
        alpha_low: u16,
        alpha_high: u16,
    },
    sudo_set_network_max_stake {
        netuid: u16,
        max_stake: u64,
    },
    sudo_set_coldkey_swap_schedule_duration {
        duration: BlockNumberFor<T>,
    },
    sudo_set_dissolve_network_schedule_duration {
        duration: BlockNumberFor<T>,
    },
    sudo_set_commit_reveal_weights_interval {
        netuid: u16,
        interval: u64,
    },
    sudo_set_evm_chain_id {
        chain_id: u64,
    },
    schedule_grandpa_change {
        next_authorities: AuthorityList,
        in_blocks: BlockNumberFor<T>,
        forced: Option<BlockNumberFor<T>>,
    },
    sudo_set_toggle_transfer {
        netuid: u16,
        toggle: bool,
    },
    sudo_toggle_evm_precompile {
        precompile_id: PrecompileEnum,
        enabled: bool,
    },
    sudo_set_subnet_moving_alpha {
        alpha: I96F32,
    },
    sudo_set_subnet_owner_hotkey {
        netuid: u16,
        hotkey: T::AccountId,
    },
    sudo_set_ema_price_halving_period {
        netuid: u16,
        ema_halving: u64,
    },
    sudo_set_alpha_sigmoid_steepness {
        netuid: u16,
        steepness: u16,
    },
    sudo_set_yuma3_enabled {
        netuid: u16,
        enabled: bool,
    },
    sudo_set_sn_owner_hotkey {
        netuid: u16,
        hotkey: T::AccountId,
    },
    sudo_set_subtoken_enabled {
        netuid: u16,
        subtoken_enabled: bool,
    },
    // some variants omitted

}

Dispatchable functions allows users to interact with the pallet and invoke state changes.

Variants

swap_authorities

The extrinsic sets the new authorities for Aura consensus. It is only callable by the root account. The extrinsic will call the Aura pallet to change the authorities.

Fields

new_authorities: BoundedVec<<T as Config>::AuthorityId, T::MaxAuthorities>

sudo_set_default_take

The extrinsic sets the default take for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the default take.

Fields

default_take: u16

sudo_set_tx_rate_limit

The extrinsic sets the transaction rate limit for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the transaction rate limit.

Fields

tx_rate_limit: u64

sudo_set_serving_rate_limit

The extrinsic sets the serving rate limit for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the serving rate limit.

Fields

netuid: u16

serving_rate_limit: u64

sudo_set_min_difficulty

The extrinsic sets the minimum difficulty for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the minimum difficulty.

Fields

netuid: u16

min_difficulty: u64

sudo_set_max_difficulty

The extrinsic sets the maximum difficulty for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the maximum difficulty.

Fields

netuid: u16

max_difficulty: u64

sudo_set_weights_version_key

The extrinsic sets the weights version key for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the weights version key.

Fields

netuid: u16

weights_version_key: u64

sudo_set_weights_set_rate_limit

The extrinsic sets the weights set rate limit for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the weights set rate limit.

Fields

netuid: u16

weights_set_rate_limit: u64

sudo_set_adjustment_interval

The extrinsic sets the adjustment interval for a subnet. It is only callable by the root account, not changeable by the subnet owner. The extrinsic will call the Subtensor pallet to set the adjustment interval.

Fields

netuid: u16

adjustment_interval: u16

sudo_set_adjustment_alpha

The extrinsic sets the adjustment alpha for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the adjustment alpha.

Fields

netuid: u16

adjustment_alpha: u64

sudo_set_max_weight_limit

The extrinsic sets the adjustment beta for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the adjustment beta.

Fields

netuid: u16

max_weight_limit: u16

sudo_set_immunity_period

The extrinsic sets the immunity period for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the immunity period.

Fields

netuid: u16

immunity_period: u16

sudo_set_min_allowed_weights

The extrinsic sets the minimum allowed weights for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the minimum allowed weights.

Fields

netuid: u16

min_allowed_weights: u16

sudo_set_max_allowed_uids

The extrinsic sets the maximum allowed UIDs for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the maximum allowed UIDs for a subnet.

Fields

netuid: u16

max_allowed_uids: u16

sudo_set_kappa

The extrinsic sets the kappa for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the kappa.

Fields

netuid: u16

kappa: u16

sudo_set_rho

The extrinsic sets the rho for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the rho.

Fields

netuid: u16

rho: u16

sudo_set_activity_cutoff

The extrinsic sets the activity cutoff for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the activity cutoff.

Fields

netuid: u16

activity_cutoff: u16

sudo_set_network_registration_allowed

The extrinsic sets the network registration allowed for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the network registration allowed.

Fields

netuid: u16

registration_allowed: bool

sudo_set_network_pow_registration_allowed

The extrinsic sets the network PoW registration allowed for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the network PoW registration allowed.

Fields

netuid: u16

registration_allowed: bool

sudo_set_target_registrations_per_interval

The extrinsic sets the target registrations per interval for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the target registrations per interval.

Fields

netuid: u16

target_registrations_per_interval: u16

sudo_set_min_burn

The extrinsic sets the minimum burn for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the minimum burn.

Fields

netuid: u16

min_burn: u64

sudo_set_max_burn

The extrinsic sets the maximum burn for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the maximum burn.

Fields

netuid: u16

max_burn: u64

sudo_set_difficulty

The extrinsic sets the difficulty for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the difficulty.

Fields

netuid: u16

difficulty: u64

sudo_set_max_allowed_validators

The extrinsic sets the maximum allowed validators for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the maximum allowed validators.

Fields

netuid: u16

max_allowed_validators: u16

sudo_set_bonds_moving_average

The extrinsic sets the bonds moving average for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the bonds moving average.

Fields

netuid: u16

bonds_moving_average: u64

sudo_set_bonds_penalty

The extrinsic sets the bonds penalty for a subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the bonds penalty.

Fields

netuid: u16

bonds_penalty: u16

sudo_set_max_registrations_per_block

The extrinsic sets the maximum registrations per block for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the maximum registrations per block.

Fields

netuid: u16

max_registrations_per_block: u16

sudo_set_subnet_owner_cut

The extrinsic sets the subnet owner cut for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the subnet owner cut.

Fields

subnet_owner_cut: u16

sudo_set_network_rate_limit

The extrinsic sets the network rate limit for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the network rate limit.

Fields

rate_limit: u64

sudo_set_tempo

The extrinsic sets the tempo for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the tempo.

Fields

netuid: u16

tempo: u16

sudo_set_total_issuance

The extrinsic sets the total issuance for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the issuance for the network.

Fields

total_issuance: u64

sudo_set_network_immunity_period

The extrinsic sets the immunity period for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the immunity period for the network.

Fields

immunity_period: u64

sudo_set_network_min_lock_cost

The extrinsic sets the min lock cost for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the min lock cost for the network.

Fields

lock_cost: u64

sudo_set_subnet_limit

The extrinsic sets the subnet limit for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the subnet limit.

Fields

max_subnets: u16

sudo_set_lock_reduction_interval

The extrinsic sets the lock reduction interval for the network. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the lock reduction interval.

Fields

interval: u64

sudo_set_rao_recycled

The extrinsic sets the recycled RAO for a subnet. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the recycled RAO.

Fields

netuid: u16

rao_recycled: u64

sudo_set_stake_threshold

The extrinsic sets the weights min stake. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the weights min stake.

Fields

min_stake: u64

sudo_set_nominator_min_required_stake

The extrinsic sets the minimum stake required for nominators. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the minimum stake required for nominators.

Fields

min_stake: u64

sudo_set_tx_delegate_take_rate_limit

The extrinsic sets the rate limit for delegate take transactions. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the rate limit for delegate take transactions.

Fields

tx_rate_limit: u64

sudo_set_min_delegate_take

The extrinsic sets the minimum delegate take. It is only callable by the root account. The extrinsic will call the Subtensor pallet to set the minimum delegate take.

Fields

take: u16

sudo_set_commit_reveal_weights_enabled

The extrinsic enabled/disables commit/reaveal for a given subnet. It is only callable by the root account or subnet owner. The extrinsic will call the Subtensor pallet to set the value.

Fields

netuid: u16

enabled: bool

sudo_set_liquid_alpha_enabled

Enables or disables Liquid Alpha for a given subnet.

Parameters

origin: The origin of the call, which must be the root account or subnet owner.

netuid: The unique identifier for the subnet.

enabled: A boolean flag to enable or disable Liquid Alpha.

Weight

This function has a fixed weight of 0 and is classified as an operational transaction that does not incur any fees.

Fields

netuid: u16

enabled: bool

sudo_set_alpha_values

Sets values for liquid alpha

Fields

netuid: u16

alpha_low: u16

alpha_high: u16

sudo_set_network_max_stake

Sets the maximum stake allowed for a specific network.

This function allows the root account to set the maximum stake for a given network. It updates the network’s maximum stake value and logs the change.

Arguments

origin - The origin of the call, which must be the root account.

netuid - The unique identifier of the network.

max_stake - The new maximum stake value to set.

Returns

Returns Ok(()) if the operation is successful, or an error if it fails.

Example

Notes

This function can only be called by the root account.

The netuid should correspond to an existing network.

TODO

Fields

netuid: u16

max_stake: u64

sudo_set_coldkey_swap_schedule_duration

Sets the duration of the coldkey swap schedule.

This extrinsic allows the root account to set the duration for the coldkey swap schedule. The coldkey swap schedule determines how long it takes for a coldkey swap operation to complete.

Arguments

origin - The origin of the call, which must be the root account.

duration - The new duration for the coldkey swap schedule, in number of blocks.

Errors

BadOrigin - If the caller is not the root account.

Weight

Weight is handled by the #[pallet::weight] attribute.

Fields

duration: BlockNumberFor<T>

sudo_set_dissolve_network_schedule_duration

Sets the duration of the dissolve network schedule.

This extrinsic allows the root account to set the duration for the dissolve network schedule. The dissolve network schedule determines how long it takes for a network dissolution operation to complete.

Arguments

origin - The origin of the call, which must be the root account.

duration - The new duration for the dissolve network schedule, in number of blocks.

Errors

BadOrigin - If the caller is not the root account.

Weight

Weight is handled by the #[pallet::weight] attribute.

Fields

duration: BlockNumberFor<T>

sudo_set_commit_reveal_weights_interval

Sets the commit-reveal weights periods for a specific subnet.

This extrinsic allows the subnet owner or root account to set the duration (in epochs) during which committed weights must be revealed. The commit-reveal mechanism ensures that users commit weights in advance and reveal them only within a specified period.

Arguments

origin - The origin of the call, which must be the subnet owner or the root account.

netuid - The unique identifier of the subnet for which the periods are being set.

periods - The number of epochs that define the commit-reveal period.

Errors

BadOrigin - If the caller is neither the subnet owner nor the root account.

SubnetDoesNotExist - If the specified subnet does not exist.

Weight

Weight is handled by the #[pallet::weight] attribute.

Fields

netuid: u16

interval: u64

sudo_set_evm_chain_id

Sets the EVM ChainID.

Arguments

origin - The origin of the call, which must be the subnet owner or the root account.

chainId - The u64 chain ID

Errors

BadOrigin - If the caller is neither the subnet owner nor the root account.

Weight

Weight is handled by the #[pallet::weight] attribute.

Fields

chain_id: u64

schedule_grandpa_change

A public interface for pallet_grandpa::Pallet::schedule_grandpa_change.

Schedule a change in the authorities.

The change will be applied at the end of execution of the block in_blocks after the current block. This value may be 0, in which case the change is applied at the end of the current block.

If the forced parameter is defined, this indicates that the current set has been synchronously determined to be offline and that after in_blocks the given change should be applied. The given block number indicates the median last finalized block number and it should be used as the canon block when starting the new grandpa voter.

No change should be signaled while any change is pending. Returns an error if a change is already pending.

Fields

next_authorities: AuthorityList

in_blocks: BlockNumberFor<T>

forced: Option<BlockNumberFor<T>>

sudo_set_toggle_transfer

Enables or disables Liquid Alpha for a given subnet.

Parameters

origin: The origin of the call, which must be the root account or subnet owner.

netuid: The unique identifier for the subnet.

enabled: A boolean flag to enable or disable Liquid Alpha.

Weight

This function has a fixed weight of 0 and is classified as an operational transaction that does not incur any fees.

Fields

netuid: u16

toggle: bool

sudo_toggle_evm_precompile

Toggles the enablement of an EVM precompile.

Arguments

origin - The origin of the call, which must be the root account.

precompile_id - The identifier of the EVM precompile to toggle.

enabled - The new enablement state of the precompile.

Errors

BadOrigin - If the caller is not the root account.

Weight

Weight is handled by the #[pallet::weight] attribute.

Fields

precompile_id: PrecompileEnum

enabled: bool

sudo_set_subnet_moving_alpha

Arguments

origin - The origin of the call, which must be the root account.

alpha - The new moving alpha value for the SubnetMovingAlpha.

Errors

BadOrigin - If the caller is not the root account.

Weight

Weight is handled by the #[pallet::weight] attribute.

Fields

alpha: I96F32

sudo_set_subnet_owner_hotkey

Change the SubnetOwnerHotkey for a given subnet.

Arguments

origin - The origin of the call, which must be the subnet owner.

netuid - The unique identifier for the subnet.

hotkey - The new hotkey for the subnet owner.

Errors

BadOrigin - If the caller is not the subnet owner or root account.

Weight

Weight is handled by the #[pallet::weight] attribute.

Fields

netuid: u16

hotkey: T::AccountId

sudo_set_ema_price_halving_period

Arguments

origin - The origin of the call, which must be the root account.

ema_alpha_period - Number of blocks for EMA price to halve

Errors

BadOrigin - If the caller is not the root account.

Weight

Weight is handled by the #[pallet::weight] attribute.

Fields

netuid: u16

ema_halving: u64

sudo_set_alpha_sigmoid_steepness

Arguments

origin - The origin of the call, which must be the root account.

netuid - The unique identifier for the subnet.

steepness - The new steepness for the alpha sigmoid function.

Errors

BadOrigin - If the caller is not the root account.

Weight

Weight is handled by the #[pallet::weight] attribute.

Fields

netuid: u16

steepness: u16

sudo_set_yuma3_enabled

Enables or disables Yuma3 for a given subnet.

Parameters

origin: The origin of the call, which must be the root account or subnet owner.

netuid: The unique identifier for the subnet.

enabled: A boolean flag to enable or disable Yuma3.

Weight

This function has a fixed weight of 0 and is classified as an operational transaction that does not incur any fees.

Fields

netuid: u16

enabled: bool

sudo_set_sn_owner_hotkey

Sets or updates the hotkey account associated with the owner of a specific subnet.

This function allows either the root origin or the current subnet owner to set or update the hotkey for a given subnet. The subnet must already exist. To prevent abuse, the call is rate-limited to once per configured interval (default: one week) per subnet.

Parameters

origin: The dispatch origin of the call. Must be either root or the current owner of the subnet.

netuid: The unique identifier of the subnet whose owner hotkey is being set.

hotkey: The new hotkey account to associate with the subnet owner.

Returns

DispatchResult: Returns Ok(()) if the hotkey was successfully set, or an appropriate error otherwise.

Errors

Error::SubnetNotExists: If the specified subnet does not exist.

Error::TxRateLimitExceeded: If the function is called more frequently than the allowed rate limit.

Access Control

Only callable by:

Root origin, or

The coldkey account that owns the subnet.

Storage

Updates [SubnetOwnerHotkey] for the given netuid.

Reads and updates [LastRateLimitedBlock] for rate-limiting.

Reads [DefaultSetSNOwnerHotkeyRateLimit] to determine the interval between allowed updates.

Rate Limiting

This function is rate-limited to one call per subnet per interval (e.g., one week).

Fields

netuid: u16

hotkey: T::AccountId

sudo_set_subtoken_enabled

Enables or disables subtoken trading for a given subnet.

Arguments

origin - The origin of the call, which must be the root account.

netuid - The unique identifier of the subnet.

subtoken_enabled - A boolean indicating whether subtoken trading should be enabled or disabled.

Errors

BadOrigin - If the caller is not the root account.

Weight

Weight is handled by the #[pallet::weight] attribute.

Fields

netuid: u16

subtoken_enabled: bool

Implementations

impl<T: Config> Call<T>

pub fn new_call_variant_swap_authorities( new_authorities: BoundedVec<<T as Config>::AuthorityId, T::MaxAuthorities>, ) -> Self

Create a call with the variant swap_authorities.

pub fn new_call_variant_sudo_set_default_take(default_take: u16) -> Self

Create a call with the variant sudo_set_default_take.

pub fn new_call_variant_sudo_set_tx_rate_limit(tx_rate_limit: u64) -> Self

Create a call with the variant sudo_set_tx_rate_limit.

pub fn new_call_variant_sudo_set_serving_rate_limit( netuid: u16, serving_rate_limit: u64, ) -> Self

Create a call with the variant sudo_set_serving_rate_limit.

pub fn new_call_variant_sudo_set_min_difficulty( netuid: u16, min_difficulty: u64, ) -> Self

Create a call with the variant sudo_set_min_difficulty.

pub fn new_call_variant_sudo_set_max_difficulty( netuid: u16, max_difficulty: u64, ) -> Self

Create a call with the variant sudo_set_max_difficulty.

pub fn new_call_variant_sudo_set_weights_version_key( netuid: u16, weights_version_key: u64, ) -> Self

Create a call with the variant sudo_set_weights_version_key.

pub fn new_call_variant_sudo_set_weights_set_rate_limit( netuid: u16, weights_set_rate_limit: u64, ) -> Self

Create a call with the variant sudo_set_weights_set_rate_limit.

pub fn new_call_variant_sudo_set_adjustment_interval( netuid: u16, adjustment_interval: u16, ) -> Self

Create a call with the variant sudo_set_adjustment_interval.

pub fn new_call_variant_sudo_set_adjustment_alpha( netuid: u16, adjustment_alpha: u64, ) -> Self

Create a call with the variant sudo_set_adjustment_alpha.

pub fn new_call_variant_sudo_set_max_weight_limit( netuid: u16, max_weight_limit: u16, ) -> Self

Create a call with the variant sudo_set_max_weight_limit.

pub fn new_call_variant_sudo_set_immunity_period( netuid: u16, immunity_period: u16, ) -> Self

Create a call with the variant sudo_set_immunity_period.

pub fn new_call_variant_sudo_set_min_allowed_weights( netuid: u16, min_allowed_weights: u16, ) -> Self

Create a call with the variant sudo_set_min_allowed_weights.

pub fn new_call_variant_sudo_set_max_allowed_uids( netuid: u16, max_allowed_uids: u16, ) -> Self

Create a call with the variant sudo_set_max_allowed_uids.

pub fn new_call_variant_sudo_set_kappa(netuid: u16, kappa: u16) -> Self

Create a call with the variant sudo_set_kappa.

pub fn new_call_variant_sudo_set_rho(netuid: u16, rho: u16) -> Self

Create a call with the variant sudo_set_rho.

pub fn new_call_variant_sudo_set_activity_cutoff( netuid: u16, activity_cutoff: u16, ) -> Self

Create a call with the variant sudo_set_activity_cutoff.

pub fn new_call_variant_sudo_set_network_registration_allowed( netuid: u16, registration_allowed: bool, ) -> Self

Create a call with the variant sudo_set_network_registration_allowed.

pub fn new_call_variant_sudo_set_network_pow_registration_allowed( netuid: u16, registration_allowed: bool, ) -> Self

Create a call with the variant sudo_set_network_pow_registration_allowed.

pub fn new_call_variant_sudo_set_target_registrations_per_interval( netuid: u16, target_registrations_per_interval: u16, ) -> Self

Create a call with the variant sudo_set_target_registrations_per_interval.

pub fn new_call_variant_sudo_set_min_burn(netuid: u16, min_burn: u64) -> Self

Create a call with the variant sudo_set_min_burn.

pub fn new_call_variant_sudo_set_max_burn(netuid: u16, max_burn: u64) -> Self

Create a call with the variant sudo_set_max_burn.

pub fn new_call_variant_sudo_set_difficulty( netuid: u16, difficulty: u64, ) -> Self

Create a call with the variant sudo_set_difficulty.

pub fn new_call_variant_sudo_set_max_allowed_validators( netuid: u16, max_allowed_validators: u16, ) -> Self

Create a call with the variant sudo_set_max_allowed_validators.

pub fn new_call_variant_sudo_set_bonds_moving_average( netuid: u16, bonds_moving_average: u64, ) -> Self

Create a call with the variant sudo_set_bonds_moving_average.

pub fn new_call_variant_sudo_set_bonds_penalty( netuid: u16, bonds_penalty: u16, ) -> Self

Create a call with the variant sudo_set_bonds_penalty.

pub fn new_call_variant_sudo_set_max_registrations_per_block( netuid: u16, max_registrations_per_block: u16, ) -> Self

Create a call with the variant sudo_set_max_registrations_per_block.

pub fn new_call_variant_sudo_set_subnet_owner_cut(subnet_owner_cut: u16) -> Self

Create a call with the variant sudo_set_subnet_owner_cut.

pub fn new_call_variant_sudo_set_network_rate_limit(rate_limit: u64) -> Self

Create a call with the variant sudo_set_network_rate_limit.

pub fn new_call_variant_sudo_set_tempo(netuid: u16, tempo: u16) -> Self

Create a call with the variant sudo_set_tempo.

pub fn new_call_variant_sudo_set_total_issuance(total_issuance: u64) -> Self

Create a call with the variant sudo_set_total_issuance.

pub fn new_call_variant_sudo_set_network_immunity_period( immunity_period: u64, ) -> Self

Create a call with the variant sudo_set_network_immunity_period.

pub fn new_call_variant_sudo_set_network_min_lock_cost(lock_cost: u64) -> Self

Create a call with the variant sudo_set_network_min_lock_cost.

pub fn new_call_variant_sudo_set_subnet_limit(max_subnets: u16) -> Self

Create a call with the variant sudo_set_subnet_limit.

pub fn new_call_variant_sudo_set_lock_reduction_interval(interval: u64) -> Self

Create a call with the variant sudo_set_lock_reduction_interval.

pub fn new_call_variant_sudo_set_rao_recycled( netuid: u16, rao_recycled: u64, ) -> Self

Create a call with the variant sudo_set_rao_recycled.

pub fn new_call_variant_sudo_set_stake_threshold(min_stake: u64) -> Self

Create a call with the variant sudo_set_stake_threshold.

pub fn new_call_variant_sudo_set_nominator_min_required_stake( min_stake: u64, ) -> Self

Create a call with the variant sudo_set_nominator_min_required_stake.

pub fn new_call_variant_sudo_set_tx_delegate_take_rate_limit( tx_rate_limit: u64, ) -> Self

Create a call with the variant sudo_set_tx_delegate_take_rate_limit.

pub fn new_call_variant_sudo_set_min_delegate_take(take: u16) -> Self

Create a call with the variant sudo_set_min_delegate_take.

pub fn new_call_variant_sudo_set_commit_reveal_weights_enabled( netuid: u16, enabled: bool, ) -> Self

Create a call with the variant sudo_set_commit_reveal_weights_enabled.

pub fn new_call_variant_sudo_set_liquid_alpha_enabled( netuid: u16, enabled: bool, ) -> Self

Create a call with the variant sudo_set_liquid_alpha_enabled.

pub fn new_call_variant_sudo_set_alpha_values( netuid: u16, alpha_low: u16, alpha_high: u16, ) -> Self

Create a call with the variant sudo_set_alpha_values.

pub fn new_call_variant_sudo_set_network_max_stake( netuid: u16, max_stake: u64, ) -> Self

Create a call with the variant sudo_set_network_max_stake.

pub fn new_call_variant_sudo_set_coldkey_swap_schedule_duration( duration: BlockNumberFor<T>, ) -> Self

Create a call with the variant sudo_set_coldkey_swap_schedule_duration.

pub fn new_call_variant_sudo_set_dissolve_network_schedule_duration( duration: BlockNumberFor<T>, ) -> Self

Create a call with the variant sudo_set_dissolve_network_schedule_duration.

pub fn new_call_variant_sudo_set_commit_reveal_weights_interval( netuid: u16, interval: u64, ) -> Self

Create a call with the variant sudo_set_commit_reveal_weights_interval.

pub fn new_call_variant_sudo_set_evm_chain_id(chain_id: u64) -> Self

Create a call with the variant sudo_set_evm_chain_id.

pub fn new_call_variant_schedule_grandpa_change( next_authorities: AuthorityList, in_blocks: BlockNumberFor<T>, forced: Option<BlockNumberFor<T>>, ) -> Self

Create a call with the variant schedule_grandpa_change.

pub fn new_call_variant_sudo_set_toggle_transfer( netuid: u16, toggle: bool, ) -> Self

Create a call with the variant sudo_set_toggle_transfer.

pub fn new_call_variant_sudo_toggle_evm_precompile( precompile_id: PrecompileEnum, enabled: bool, ) -> Self

Create a call with the variant sudo_toggle_evm_precompile.

pub fn new_call_variant_sudo_set_subnet_moving_alpha(alpha: I96F32) -> Self

Create a call with the variant sudo_set_subnet_moving_alpha.

pub fn new_call_variant_sudo_set_subnet_owner_hotkey( netuid: u16, hotkey: T::AccountId, ) -> Self

Create a call with the variant sudo_set_subnet_owner_hotkey.

pub fn new_call_variant_sudo_set_ema_price_halving_period( netuid: u16, ema_halving: u64, ) -> Self

Create a call with the variant sudo_set_ema_price_halving_period.

pub fn new_call_variant_sudo_set_alpha_sigmoid_steepness( netuid: u16, steepness: u16, ) -> Self

Create a call with the variant sudo_set_alpha_sigmoid_steepness.

pub fn new_call_variant_sudo_set_yuma3_enabled( netuid: u16, enabled: bool, ) -> Self

Create a call with the variant sudo_set_yuma3_enabled.

pub fn new_call_variant_sudo_set_sn_owner_hotkey( netuid: u16, hotkey: T::AccountId, ) -> Self

Create a call with the variant sudo_set_sn_owner_hotkey.

pub fn new_call_variant_sudo_set_subtoken_enabled( netuid: u16, subtoken_enabled: bool, ) -> Self

Create a call with the variant sudo_set_subtoken_enabled.