pallet_admin_utils::pallet

Trait Config

pub trait Config:
    Config
    + Config
    + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Aura: AuraInterface<<Self as Config>::AuthorityId, Self::MaxAuthorities>;
    type Grandpa: GrandpaInterface<Self>;
    type AuthorityId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize + MaxEncodedLen;
    type MaxAuthorities: Get<u32>;
    type Balance: Balance;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Aura: AuraInterface<<Self as Config>::AuthorityId, Self::MaxAuthorities>

Implementation of the AuraInterface

type Grandpa: GrandpaInterface<Self>

Implementation of GrandpaInterface

type AuthorityId: Member + Parameter + RuntimeAppPublic + MaybeSerializeDeserialize + MaxEncodedLen

The identifier type for an authority.

type MaxAuthorities: Get<u32>

The maximum number of authorities that the pallet can hold.

type Balance: Balance

Unit of assets

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.